+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 ## Changed
+- Renamed `kennethormandy\marketplace\services\AccountsService` → `kennethormandy\marketplace\services\Accounts`
+- Renamed `kennethormandy\marketplace\services\FeesService` → `kennethormandy\marketplace\services\Fees`
+- Renamed `kennethormandy\marketplace\services\HandlesService` → `kennethormandy\marketplace\services\Handles`
+- Renamed `kennethormandy\marketplace\services\PayeesService` → `kennethormandy\marketplace\services\Payees`
 - Renamed `handlesService` to `handles`
 
 ## Removed